requirements.txt
Flask==2.2.2
requests==2.28.1
Flask_SQLAlchemy==2.5.1
poetry==1.2.1
Flask-WTF==1.0.1
WTForms==3.0.1